ref(discover) Move functions only used by discover1 out of general utils
Move a few discover1 functions out of the utils/snuba and into modules
that are proximal to the endpoints that use them. I've moved `zerofill`
into discover2 code as that is a better long term home for it with
discover1 going away in 'the future'.
