+++ src/sentry/discover/endpoints/discover_query.py
@@ -10,6 +10,7 @@ from sentry.api.bases import OrganizationEndpoint
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.paginator import GenericOffsetPaginator
 from sentry.utils import snuba
+from sentry.discover.utils import transform_aliases_and_query
 from sentry import features
 
 from .serializers import DiscoverQuerySerializer
@@ -83,7 +84,7 @@ class DiscoverQueryEndpoint(OrganizationEndpoint):
 
         if not kwargs["aggregations"]:
 
-            data_fn = partial(snuba.transform_aliases_and_query, referrer="discover", **kwargs)
+            data_fn = partial(transform_aliases_and_query, referrer="discover", **kwargs)
             return self.paginate(
                 request=request,
                 on_results=lambda results: self.handle_results(results, requested_query, projects),
@@ -91,7 +92,7 @@ class DiscoverQueryEndpoint(OrganizationEndpoint):
                 max_per_page=1000,
             )
         else:
-            snuba_results = snuba.transform_aliases_and_query(referrer="discover", **kwargs)
+            snuba_results = transform_aliases_and_query(referrer="discover", **kwargs)
             return Response(
                 self.handle_results(snuba_results, requested_query, projects), status=200
             )
